# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: EHR API
  description: |
    Headless Electronic Health Record API with dual interfaces:
    - FHIR R4-compliant endpoints under /fhir
    - Operational REST endpoints under /api/v1
  version: 0.1.0
  contact:
    name: EHR Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /fhir/Observation:
    get:
      tags: [Observation]
      summary: Search observations (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - name: category
          in: query
          schema:
            type: string
          description: Observation category (e.g. vital-signs, laboratory)
        - name: code
          in: query
          schema:
            type: string
          description: LOINC or other code
        - $ref: '#/components/parameters/FHIRDate'
        - $ref: '#/components/parameters/FHIRStatus'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Observation]
      summary: Create observation (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Observation'
      responses:
        '201':
          description: Observation created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Observation'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/Condition:
    get:
      tags: [Condition]
      summary: Search conditions (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - name: clinical-status
          in: query
          schema:
            type: string
          description: Clinical status (active, recurrence, relapse, inactive, remission, resolved)
        - name: category
          in: query
          schema:
            type: string
          description: Condition category
        - name: code
          in: query
          schema:
            type: string
          description: ICD-10 or SNOMED code
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Condition]
      summary: Create condition (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Condition'
      responses:
        '201':
          description: Condition created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Condition'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/AllergyIntolerance:
    get:
      tags: [AllergyIntolerance]
      summary: Search allergies (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - name: clinical-status
          in: query
          schema:
            type: string
          description: Clinical status (active, inactive, resolved)
        - name: type
          in: query
          schema:
            type: string
            enum: [allergy, intolerance]
        - name: criticality
          in: query
          schema:
            type: string
            enum: [low, high, unable-to-assess]
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [AllergyIntolerance]
      summary: Create allergy (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/AllergyIntolerance'
      responses:
        '201':
          description: AllergyIntolerance created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/AllergyIntolerance'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/Procedure:
    get:
      tags: [Procedure]
      summary: Search procedures (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: code
          in: query
          schema:
            type: string
          description: CPT or SNOMED code
        - $ref: '#/components/parameters/FHIRDate'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Procedure]
      summary: Create procedure (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Procedure'
      responses:
        '201':
          description: Procedure created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Procedure'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/observations:
    get:
      tags: [Observation]
      summary: List observations
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of observations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Observation]
      summary: Create observation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationInput'
      responses:
        '201':
          description: Observation created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/observations/{id}:
    get:
      tags: [Observation]
      summary: Get observation by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Observation details
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Observation]
      summary: Update observation
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Observation updated
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Observation]
      summary: Delete observation
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Observation deleted

  # ── Conditions (Operational) ───────────────────────────────────────
  /api/v1/conditions:
    get:
      tags: [Condition]
      summary: List conditions
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of conditions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Condition]
      summary: Create condition
      responses:
        '201':
          description: Condition created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/conditions/{id}:
    get:
      tags: [Condition]
      summary: Get condition by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Condition details
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Condition]
      summary: Update condition
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Condition updated
    delete:
      tags: [Condition]
      summary: Delete condition
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Condition deleted

  # ── Medication Requests (Operational) ──────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. In development mode, the server accepts any
        well-formed JWT. In production, tokens are validated against
        the configured JWKS endpoint.
  parameters:
...
    FHIRPatient:
      name: patient
      in: query
      schema:
        type: string
      description: Patient reference (e.g. Patient/123)
    FHIRStatus:
      name: status
      in: query
      schema:
        type: string
      description: Resource status code
    FHIRDate:
      name: date
      in: query
      schema:
        type: string
      description: Date search (supports prefixes eq, ne, gt, lt, ge, le, sa, eb)
...
    Count:
      name: _count
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page (FHIR)
    Offset:
      name: _offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination (FHIR)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination
  responses:
    FHIRBundle:
      description: FHIR Bundle (searchset)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/Bundle'
    BadRequest:
      description: Invalid request
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: missing or invalid token
...
    InternalError:
      description: Internal server error
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
  schemas:
...
    Observation:
      type: object
      required: [resourceType, status, code]
      properties:
        resourceType:
          type: string
          enum: [Observation]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [registered, preliminary, final, amended, corrected, cancelled, entered-in-error, unknown]
        category:
          type: array
          items:
            $ref: '#/components/schemas/CodeableConcept'
        code:
          $ref: '#/components/schemas/CodeableConcept'
        subject:
          $ref: '#/components/schemas/Reference'
        encounter:
          $ref: '#/components/schemas/Reference'
        effectiveDateTime:
          type: string
          format: date-time
        valueQuantity:
          $ref: '#/components/schemas/Quantity'
        valueString:
          type: string
        component:
          type: array
          items:
            type: object
            properties:
              code:
                $ref: '#/components/schemas/CodeableConcept'
              valueQuantity:
                $ref: '#/components/schemas/Quantity'
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          description: Total number of matching records
        limit:
          type: integer
        offset:
          type: integer
...